+++ Instructors/views/multipleChoiceView.py
@@ -12,6 +12,7 @@ from Instructors.models import Challenges, ChallengesQuestions
 from Instructors.views.utils import initialContextDict, getCourseSkills, addSkillsToQuestion, saveTags, getSkillsForQuestion, extractTags
 from Badges.enums import QuestionTypes, ObjectTypes
 from Instructors.constants import unassigned_problems_challenge_name
+from decimal import Decimal
 
 from django.contrib.auth.decorators import login_required
 
@@ -136,7 +137,7 @@ def multipleChoiceForm(request):
                 
             challengeID = request.POST['challengeID']
             challenge = Challenges.objects.get(pk=int(challengeID))
-            ChallengesQuestions.addQuestionToChallenge(question, challenge, int(request.POST['points']), position)
+            ChallengesQuestions.addQuestionToChallenge(question, challenge, Decimal(request.POST['points']), position)
 
            
             addSkillsToQuestion(currentCourse,question,request.POST.getlist('skills[]'),request.POST.getlist('skillPoints[]'))
